chore(ci): Add --load flag to docker buildx command
A slight change has been made to the docker buildx command in the build script. The --load flag has been added in order to specify that the built image should be loaded into Docker as part of the build process. This could potentially increase the efficiency of subsequent Docker commands that use the same image.